V0-add alert delete

+++ src/app/components/assunto/assunto-list/assunto-list.component.ts
@@ -44,21 +44,22 @@ export class AssuntoListComponent implements OnInit {
   }
 
   delete(id: any): void {
-    this.service.delete(id).subscribe(() => {
-      this.toast.success('Assunto excluído com sucesso', 'Delete');
-      this.findAll();
-      }, ex => {
-      console.log(ex);
-        if(ex.error.errors) {
-          ex.error.errors.forEach(element => {
-            this.toast.error(element.message);          
-          });
-        } else {
-          this.toast.error(ex.error.message);
-        }      
-      }
-    )
-  }  
-
+    if (confirm('Tem certeza que deseja excluir este assunto?')) {
+      this.service.delete(id).subscribe(() => {
+        this.toast.success('Assunto excluído com sucesso', 'Delete');
+        this.findAll();
+        }, ex => {
+        console.log(ex);
+          if(ex.error.errors) {
+            ex.error.errors.forEach(element => {
+              this.toast.error(element.message);          
+            });
+          } else {
+            this.toast.error(ex.error.message);
+          }      
+        }
+      )
+    }
+  }
 }
